feat(lint): noNoninteractiveTabindex
