+++ internal/compiler/lint/rules/a11y/noNoninteractiveTabindex.test.md
@@ -2,18 +2,17 @@
 
 **DO NOT MODIFY**. This file has been autogenerated. Run `rome test internal/compiler/lint/rules/harness.test.ts --update-snapshots` to update.
 
-## `jsx-a11y/noNoninteractiveTabindex`
+## `a11y/noNoninteractiveTabindex`
 
 ### `0`
 
 ```
 
- lint/jsx-a11y/noNoninteractiveTabindex/reject/1/file.tsx:1:5 lint/jsx-a11y/noNoninteractiveTabindex
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+ lint/a11y/noNoninteractiveTabindex/reject/1/file.tsx:1:5 lint/a11y/noNoninteractiveTabindex ━━━━━━━
 
   ✖ Do not use tabIndex on an element that is not interactive.
 
-    <div tabIndex='0' ></div>
+    <div tabIndex="0" ></div>
          ^^^^^^^^^^^^
 
   ℹ Adding non-interactive elements to the keyboard navigation flow can confuse users.
@@ -35,12 +34,11 @@
 
 ```
 
- lint/jsx-a11y/noNoninteractiveTabindex/reject/2/file.tsx:1:5 lint/jsx-a11y/noNoninteractiveTabindex
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+ lint/a11y/noNoninteractiveTabindex/reject/2/file.tsx:1:5 lint/a11y/noNoninteractiveTabindex ━━━━━━━
 
   ✖ Do not use tabIndex on an element that is not interactive.
 
-    <div role='article' tabIndex='0' ></div>
+    <div role="article" tabIndex="0" ></div>
          ^^^^^^^^^^^^^^
 
   ℹ Adding non-interactive elements to the keyboard navigation flow can confuse users.
@@ -62,8 +60,7 @@
 
 ```
 
- lint/jsx-a11y/noNoninteractiveTabindex/reject/3/file.tsx:1:9 lint/jsx-a11y/noNoninteractiveTabindex
-  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+ lint/a11y/noNoninteractiveTabindex/reject/3/file.tsx:1:9 lint/a11y/noNoninteractiveTabindex ━━━━━━━
 
   ✖ Do not use tabIndex on an element that is not interactive.
 
@@ -221,6 +218,165 @@
 ### `12: formatted`
 
 ```tsx
-<button tabindex="-1" />;
+<button tabIndex="-1" />;
+
+```
+
+### `13`
+
+```
+
+ lint/a11y/noNoninteractiveTabindex/reject/1/file.html:1:5 lint/a11y/noNoninteractiveTabindex ━━━━━━
+
+  ✖ Do not use tabIndex on an element that is not interactive.
+
+    <div tabindex="0" ></div>
+         ^^^^^^^^^^^^
+
+  ℹ Adding non-interactive elements to the keyboard navigation flow can confuse users.
+
+━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+
+✖ Found 1 problem
+
+```
+
+### `13: formatted`
+
+```html
+<div tabindex="0">
+</div>
+
+```
+
+### `14`
+
+```
+
+ lint/a11y/noNoninteractiveTabindex/reject/2/file.html:1:5 lint/a11y/noNoninteractiveTabindex ━━━━━━
+
+  ✖ Do not use tabIndex on an element that is not interactive.
+
+    <div role="article" tabindex="0" ></div>
+         ^^^^^^^^^^^^^^
+
+  ℹ Adding non-interactive elements to the keyboard navigation flow can confuse users.
+
+━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+
+✖ Found 1 problem
+
+```
+
+### `14: formatted`
+
+```html
+<div role="article" tabindex="0">
+</div>
+
+```
+
+### `15`
+
+```
+✔ No known problems!
+
+```
+
+### `15: formatted`
+
+```html
+<button tabindex="0">
+</button>
+
+```
+
+### `16`
+
+```
+✔ No known problems!
+
+```
+
+### `16: formatted`
+
+```html
+<span role="button" tabindex="0">
+</span>
+
+```
+
+### `17`
+
+```
+✔ No known problems!
+
+```
+
+### `17: formatted`
+
+```html
+<span role="article" tabindex="-1">
+</span>
+
+```
+
+### `18`
+
+```
+✔ No known problems!
+
+```
+
+### `18: formatted`
+
+```html
+<article tabindex="-1">
+</article>
+
+```
+
+### `19`
+
+```
+✔ No known problems!
+
+```
+
+### `19: formatted`
+
+```html
+<div tabindex="-1">
+</div>
+
+```
+
+### `20`
+
+```
+✔ No known problems!
+
+```
+
+### `20: formatted`
+
+```html
+<div>
+</div>
+
+```
+
+### `21`
+
+```
+✔ No known problems!
+
+```
+
+### `21: formatted`
+
+```html
+<button tabindex="-1">
+</button>
 
 ```
