--- internal/compiler/lint/rules/jsx-a11y/noNoninteractiveTabindex.test.rjson
@@ -1,18 +1,0 @@
-filename: "file.tsx"
-invalid: [
-	"<div tabIndex='0' ></div>"
-	"<div role='article' tabIndex='0' ></div>"
-	"<article tabIndex={0} />"
-]
-valid: [
-	"<button tabIndex='0' ></button>"
-	"<span role='button' tabIndex='0' ></span>"
-	"<span role='article' tabIndex='-1' ></span>"
-	"<MyButton tabIndex={0} />"
-	"<article tabIndex='-1'></article>"
-	"<div tabIndex='-1' ></div>"
-	"<article tabIndex={-1}></article>"
-	"<div tabIndex={-1} ></div>"
-	"<div></div>"
-	"<button tabindex='-1'></button>"
-]
