+++ internal/compiler/lint/rules/tests.ts
@@ -8,13 +8,15 @@ type Test = {
 
 type Tests = Dict<Test | (Test[]) | (Test[][])>;
 
-/* GENERATED:START(hash:02b7feddd7b65e88b8e890ff7b6238ed22ffcbe9,id:main) Everything below is automatically generated. DO NOT MODIFY. Run `./rome run scripts/generated-files/lint-rules` to update. */
+/* GENERATED:START(hash:62ad21b6efdfcda32aa6185b623728dc0e5b14eb,id:main) Everything below is automatically generated. DO NOT MODIFY. Run `./rome run scripts/generated-files/lint-rules` to update. */
 // @ts-ignore
 import noAriaUnsupportedElements from "./a11y/noAriaUnsupportedElements.test.rjson";
 // @ts-ignore
 import noDistractingElements from "./a11y/noDistractingElements.test.rjson";
 // @ts-ignore
 import noNoninteractiveElementToInteractiveRole from "./a11y/noNoninteractiveElementToInteractiveRole.test.rjson";
+// @ts-ignore
+import noNoninteractiveTabindex from "./a11y/noNoninteractiveTabindex.test.rjson";
 // @ts-ignore
 import noSvgWithoutTitle from "./a11y/noSvgWithoutTitle.test.rjson";
 // @ts-ignore
@@ -130,8 +132,6 @@ import noAutofocus from "./jsx-a11y/noAutofocus.test.rjson";
 // @ts-ignore
 import noHeaderScope from "./jsx-a11y/noHeaderScope.test.rjson";
 // @ts-ignore
-import noNoninteractiveTabindex from "./jsx-a11y/noNoninteractiveTabindex.test.rjson";
-// @ts-ignore
 import noOnChange from "./jsx-a11y/noOnChange.test.rjson";
 // @ts-ignore
 import noPositiveTabindex from "./jsx-a11y/noPositiveTabindex.test.rjson";
@@ -234,6 +234,7 @@ export const tests: Tests = {
 	"a11y/noAriaUnsupportedElements": noAriaUnsupportedElements,
 	"a11y/noDistractingElements": noDistractingElements,
 	"a11y/noNoninteractiveElementToInteractiveRole": noNoninteractiveElementToInteractiveRole,
+	"a11y/noNoninteractiveTabindex": noNoninteractiveTabindex,
 	"a11y/noSvgWithoutTitle": noSvgWithoutTitle,
 	"a11y/useAltText": useAltText,
 	"a11y/useAriaProptypes": useAriaProptypes,
@@ -291,7 +292,6 @@ export const tests: Tests = {
 	"jsx-a11y/noAccessKey": noAccessKey,
 	"jsx-a11y/noAutofocus": noAutofocus,
 	"jsx-a11y/noHeaderScope": noHeaderScope,
-	"jsx-a11y/noNoninteractiveTabindex": noNoninteractiveTabindex,
 	"jsx-a11y/noOnChange": noOnChange,
 	"jsx-a11y/noPositiveTabindex": noPositiveTabindex,
 	"jsx-a11y/noRedundantAlt": noRedundantAlt,
