+++ internal/diagnostics/categories.ts
@@ -96,11 +96,12 @@ export type DiagnosticCategoryPrefix =
 	| "typeCheck"
 	| "v8";
 
-/* GENERATED:START(hash:fa7103cf2c9efd4145aa2b4ac4938af65c733b8d,id:main) Everything below is automatically generated. DO NOT MODIFY. Run `./rome run scripts/generated-files/lint-rules` to update. */
+/* GENERATED:START(hash:65248b89544465d372d5a5b78469b1c2a6239793,id:main) Everything below is automatically generated. DO NOT MODIFY. Run `./rome run scripts/generated-files/lint-rules` to update. */
 export type DiagnosticLintCategory =
 	| "lint/a11y/noAriaUnsupportedElements"
 	| "lint/a11y/noDistractingElements"
 	| "lint/a11y/noNoninteractiveElementToInteractiveRole"
+	| "lint/a11y/noNoninteractiveTabindex"
 	| "lint/a11y/noSvgWithoutTitle"
 	| "lint/a11y/useAltText"
 	| "lint/a11y/useAriaProptypes"
@@ -159,7 +160,6 @@ export type DiagnosticLintCategory =
 	| "lint/jsx-a11y/noAccessKey"
 	| "lint/jsx-a11y/noAutofocus"
 	| "lint/jsx-a11y/noHeaderScope"
-	| "lint/jsx-a11y/noNoninteractiveTabindex"
 	| "lint/jsx-a11y/noOnChange"
 	| "lint/jsx-a11y/noPositiveTabindex"
 	| "lint/jsx-a11y/noRedundantAlt"
@@ -217,6 +217,7 @@ const lintCategoryNameMap: {[name in DiagnosticLintCategory]: true} = {
 	"lint/a11y/noAriaUnsupportedElements": true,
 	"lint/a11y/noDistractingElements": true,
 	"lint/a11y/noNoninteractiveElementToInteractiveRole": true,
+	"lint/a11y/noNoninteractiveTabindex": true,
 	"lint/a11y/noSvgWithoutTitle": true,
 	"lint/a11y/useAltText": true,
 	"lint/a11y/useAriaProptypes": true,
@@ -275,7 +276,6 @@ const lintCategoryNameMap: {[name in DiagnosticLintCategory]: true} = {
 	"lint/jsx-a11y/noAccessKey": true,
 	"lint/jsx-a11y/noAutofocus": true,
 	"lint/jsx-a11y/noHeaderScope": true,
-	"lint/jsx-a11y/noNoninteractiveTabindex": true,
 	"lint/jsx-a11y/noOnChange": true,
 	"lint/jsx-a11y/noPositiveTabindex": true,
 	"lint/jsx-a11y/noRedundantAlt": true,
